fix serialization and tests (such that the transmission tests use ValidatedEvents)
